[shivanshs9] Changed section structure
